+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added `mode` parameter to `DBNStore.to_csv` to control the file writing mode
 - Added `mode` parameter to `DBNStore.to_json` to control the file writing mode
 - Added `mode` parameter to `DBNStore.to_parquet` to control the file writing mode
+- Added `compression` parameter to `DBNStore.to_file` which controls the output compression format
 
 #### Breaking changes
 - Changed default write mode for `DBNStore.to_csv` to overwrite ("w")